// src/app/app-shared/models/student.ts
export class Student {
    id?: number;
    name: string;
    category: string;
    dateOfBirth: Date;
    domicileCertificate: boolean;
    birthCertificate: boolean;
    previousMarkSheet: boolean;
    policeClearance: boolean;
    passport: boolean;
    signedDeclaration: boolean;
    fatherName: string;
    motherName: string;
    lastClassScore: number;
}